update
update

+++ backend/App/controllers/jewels-controller.js
@@ -4,36 +4,6 @@ const {validationResult} = require('express-validator')
 const _ = require('lodash')
 const jewelsCltr = {}
 
-// jewelsCltr.create = async (req, res) => {
-//     const errors = validationResult(req)
-//     if(!errors.isEmpty()){
-//         return res.status(404).json({errors:errors.array()})
-//     }
-//     try {
-//         console.log(req.files)
-//         const result =[]
-//         req.files.forEach(ele =>{
-//             result.push(ele.filename)
-//         })
-        
-//       const images = req.files.map(file => file.path)
-//       console.log(images)
-//       const body = _.pick(req.body,['image','price','caption'])
-//       const shop = await Shop.findOne({id:req.user.id})
-//       body.shopId = shop.id
-//       const jewel = new Jewels(body)
-//       jewel.images = result
-//       console.log(jewel)
-//       const savedJewel = await jewel.save()
-  
-//       res.status(200).json({ message: 'Files uploaded successfully!', jewel: savedJewel })
-//     } catch (error) {
-//       console.error(error)
-//       res.status(500).json({ error: 'Internal Server Error' })
-//     }
-//   }
-
-
 jewelsCltr.create = async (req, res) => {
     const errors = validationResult(req)
     if (!errors.isEmpty()) {
@@ -42,17 +12,16 @@ jewelsCltr.create = async (req, res) => {
 
     try {
         const { body, file } = req
-
-        // const id = req.id
-        
-        // const shopOwner = await Shop.findById(id)
-        // if (!shopOwner) {
-        //     return res.status(404).json({ error: 'Shop owner not found' })
-        // }
+        const owner = req.user.id
+        const shop = await Shop.find({ownerId:owner})        
+        if (!shop) {
+            return res.status(404).json({ error: 'Shop owner not found' })
+        }
 
         const jewel = new Jewels({
             ...body,
-            // owner: id,
+            ownerId: owner,
+            shopId : shop._id,
             images: file.path 
         })
 
@@ -75,16 +44,17 @@ jewelsCltr.get = async(req,res) =>{
         res.status(500).json({error:'Internal Server Error'})
     }
 }
-jewelsCltr.get = async(req,res) =>{
-    try{
-        const jewel = await Jewels.find()
-        res.json(jewel)
-    }
-    catch(err){
-        console.log(err)
-        res.status(500).json({error:'Internal Server Error'})
-    }
-}
+
+// jewelsCltr.getAll = async(req,res) =>{
+//     try{
+//         const jewel = await Jewels.find()
+//         res.json(jewel)
+//     }
+//     catch(err){
+//         console.log(err)
+//         res.status(500).json({error:'Internal Server Error'})
+//     }
+// }
 
 jewelsCltr.update = async(req,res)=>{
     const errors = validationResult(req)
@@ -94,9 +64,8 @@ jewelsCltr.update = async(req,res)=>{
     try{
         const id = req.params.id
         const body = _.pick(req.body,['images','price','caption'])
-        // const shop = await Shop.findOne({id:req.user.id})
-        // body.id = shop.id
-        const jewel = await Jewels.findOneAndUpdate({_id:id},body,{new:true})
+        // const shop = await Shop.findOne({ownerId:req.user.id})
+        const jewel = await Jewels.findOneAndUpdate({_id:id},{ownerId:req.user.id},body,{new:true})
         res.json(jewel)
     }
     catch(err){
@@ -108,7 +77,7 @@ jewelsCltr.update = async(req,res)=>{
 jewelsCltr.delete = async(req,res) => {
     try{
         const id = req.params.id
-        const jewel = await Jewels.findOneAndDelete({_id:id})
+        const jewel = await Jewels.findOneAndDelete({_id:id},{ownerId:req.user.id})
         res.json(jewel)
     }
     catch(err){
